Compilation script now only generates a.out if compilation is successful.

+++ src/compile.sh
@@ -1,8 +1,10 @@
 #!/usr/bin/env bash
-gcc -c putting_it_all_together/runtime.c
 
 python3 main.py "$1"
-gcc -no-pie -g output.s runtime.o
+if [ $? -eq 0 ]; then
+  gcc -c putting_it_all_together/runtime.c
+  gcc -no-pie -g output.s runtime.o
 
-rm "output.s"
-rm "runtime.o"
+  rm "output.s"
+  rm "runtime.o"
+fi
